+++ packages/skill/eslint.config.js
@@ -1,0 +1,3 @@
+import { config } from "@weekly-git-report/eslint-config/base";
+
+export default config;